+++ NEWS.md
@@ -300,6 +300,7 @@ ion for time-series (#12670)
 * Fix histogram zoom (#12945)
 * Fix ambiguous column call in the search tweets query (#13073)
 * Fix email validator failing with empty emails (#13078)
+* Be sure to delete the analysis cache tables while we're dropping a organization user (#13136)
 * Treat all time series dataview timestamps as UTC (#13070)
 
 ### Internals